// api-stage/server/src/common/enums/user-role.enum.ts
export enum UserRole {
    SUPER_ADMIN = 'SUPER_ADMIN',
    TENANT_ADMIN = 'TENANT_ADMIN',
    ADMIN = 'ADMIN',
    USER = 'USER',
    VIEWER = 'VIEWER',
    SUPPORT = 'SUPPORT',
    MANAGER = 'MANAGER',
    TECHNICIAN = 'TECHNICIAN',
    WORKSHOP_MANAGER = 'WORKSHOP_MANAGER',
    RECEPTION = 'RECEPTION',
    SERVICE_MANAGER = 'SERVICE_MANAGER',
    PROCUREMENT = 'PROCUREMENT',
    WAREHOUSE = 'WAREHOUSE',
    ADVISOR = 'ADVISOR',
    PARTS_MANAGER = 'PARTS_MANAGER'
}